==> vignettes/1_main_text.Rmd <==
---
title: "1 main text"
author: "Ruoyu Wu, Yiluan Song"
date: "2025-04-30"
output: rmarkdown::html_vignette
vignette: >
  %\VignetteIndexEntry{1 main text}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE, process=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache = FALSE, warning = FALSE, message = FALSE, process = FALSE)

library(phenologyspore)

.fulldata <- F
.datsave <- F
.figsave <- T

if (exists(".fulldata") && isTRUE(.fulldata)) {
  .path <- list(
    input = "../alldata/input/",
    intermediate = "../alldata/intermediate/",
    output = "../alldata/output/"
  )
} else {
  .path <- list(
    output = "../inst/"
  )
}
```

# Spore phenology metrics
```{r fig-metrics, fig.height=11, fig.width=11, fig.cap="Figure 1. Definition and numerical summary of ten fungal spore metrics. (a) The annotation of ten fungal spore metrics. Orange indicates metrics defined in the ecological approach, while dark red indicates metrics defined in the public health approach. Each box represents each metric’s interquartile range (IQR), with the median value indicated by the line inside. Whiskers extend to the minimum and maximum values within 1.5 times the IQR, while points beyond the whiskers are considered outliers. (c) Phenology metrics are in the unit of day of spore year (DOY) or days (for season lengths LOS, LAS), while intensity metrics are in the unit of (b) spores m–3 or (d) spores m–3 days. The abbreviations of metrics used in (b), (c), and (d) are defined in (a)."}
if (.fulldata) {
  df_sample <- tidy_sample_data(
    fill_smooth_offset_path = str_c(.path$intermediate, "fill_smooth_offset.rds"),
    metrics_amplitude_path = str_c(.path$intermediate, "metrics_amplitude.rds")
  )
  if (.datsave) {
    write_rds(df_sample, str_c(.path$intermediate, "sample.rds"))
    usethis::use_data(df_sample, overwrite = T)
  }
}

(p_conceptual <- plot_conceptual(df_sample, df_analysis))

if (.figsave) {
  ggsave(
    str_c(.path$output, "figures/Figure_1.png"),
    plot = p_conceptual,
    height = 11, width = 11,
    device = png, type = "cairo"
  )
}
```

# Spore phenology calendar
```{r fig-calendar, fig.width=10, fig.height=5*0.8, fig.cap="Figure 2. Fungal spore calendar for ten sampling stations. The daily long-term mean of fungal spore concentration is displayed, 2003-2022. Darker colors indicate higher concentrations while missing data are represented in white. (a) Stations show variable seasonality across ecological regions. (b) Stations with higher long-term mean of total annual precipitation (mm) show higher intensity of fungal spore seasons."}
if (.fulldata) {
  df_ts <- read_rds(str_c(.path$intermediate, "ts.rds"))
  (p_calendar <- plot_calendar(df_ts, df_daymet_annual))

  if (.figsave) {
    ggsave(
      str_c(.path$output, "figures/Figure_2.png"),
      plot = p_calendar,
      height = 5 * 0.8, width = 10,
      device = png, type = "cairo"
    )
  }
} else {
  (p_calendar <- cowplot::ggdraw() +
    cowplot::draw_image(str_c(.path$output, "figures/Figure_2.png")))
}
```

# Station level trend
```{r fig-trend-station, fig.height=9*0.618, fig.width=9, fig.cap="Figure 3. Station-level temporal trends of the start of allergy season (SAS). The trends were estimated by Theil-Sen linear regression. Warm colors indicate earlier seasons, and circle sizes are proportional to the years of data at each station."}
df_trend_station_SAS <- calc_trend_station_TS(df_in = df_analysis, ls_metric = "SAS", pct = 0.8)
(p_trend_map_TS_SAS <- plot_trend_map_TS(df_in = df_trend_station_SAS, ls_metric = "SAS"))

if (.figsave) {
  ggsave(
    str_c(.path$output, "figures/Figure_3.png"),
    plot = p_trend_map_TS_SAS,
    height = 9 * 0.618, width = 9,
    device = png, type = "cairo"
  )
}
```

# Continent level trend
```{r fig-trend-ctnt, fig.width=11, fig.height=11*0.8, fig.cap="Figure 4. Temporal trends of ten fungal spore metrics across all the stations. The estimated trends from linear mixed effects models across stations of the year against (a) the start of season (SOS), (f) start of allergy season (SAS), (b) end of season (EOS), (g) end of allergy season (EAS), (c) length of season (LOS), (h) length of allergy season (LAS), (d) amplitude concentration (Ca), (i) peak concentration (Cp), (e) annual integral (AIn), and (j) allergy season integral (ASIn). Solid black lines indicate significant (p < 0.05, t-test) slopes. Shaded areas indicate the 95% confidence intervals of the fixed effect. Points are individual years at individual stations. Colors are station-level Theil-Sen linear regression slopes, with warmer colors indicating earlier days of year, longer days, or higher intensity. The slope of colored lines are model-predicted station-level trends. The top row presents metrics defined in the ecological approach, while the bottom row shows metrics defined in the public health approach. Intensity metrics are transformed using the natural logarithm. To highlight trends, all intensity metrics are displayed on a natural logarithm scale, consistent with the transformations used during model fitting."}
(p_trend_ctnt_pred <- plot_trend_ctnt_pred(df_analysis))

if (.figsave) {
  ggsave(
    str_c(.path$output, "figures/Figure_4.png"),
    plot = p_trend_ctnt_pred,
    height = 11 * 0.8, width = 11,
    device = png, type = "cairo"
  )
}
```

# Coefficient plot
```{r fig-coefplot, fig.height=10*0.618, fig.width=10, fig.cap="Figure 5. Summary of coefficients β1 in three mixed-effects models for each metric. Points are estimated values of β1. Error bars indicate the 95% confidence intervals of the fixed slope. Specific values are annotated. Colors are different independent variables in the model. Non-transparent colors present significant coefficients (p < 0.05, t-test)."}
df_lme <- calc_lme_all(df_analysis, x_vrb = c("year_new", "MAT", "TAP"), pct = 0.8)
if (.datsave) {
  write_rds(df_lme, str_c(.path$intermediate, "lme.rds"))
  usethis::use_data(df_lme, overwrite = T)
}

(p_coef <- plot_coef(df_lme))

if (.figsave) {
  ggsave(
    str_c(.path$output, "figures/Figure_5.png"),
    plot = p_coef,
    height = 10 * 0.618, width = 10,
    device = png, type = "cairo"
  )
}
```
